reorganisation and clean
